+++ wallet/Views/CheddarActionBar.swift
@@ -31,7 +31,7 @@ class CheddarActionBar: UIView {
         
         // Add shadow
         layer.shadowColor = Theme.shadowColor.cgColor
-        layer.shadowOffset = CGSize(width: 0.0, height: -CGFloat(Dimens.shadow))
+        layer.shadowOffset = CGSize(width: 0.0, height: -Dimens.shadow)
         layer.shadowRadius = 0.0
         layer.shadowOpacity = 1.0
         layer.masksToBounds = false
@@ -44,10 +44,10 @@ class CheddarActionBar: UIView {
         leftButton!.title = title
         addSubview(leftButton!)
         leftButton!.translatesAutoresizingMaskIntoConstraints = false
-        leftButton!.topAnchor.constraint(equalTo: topAnchor, constant: CGFloat(Dimens.mediumMargin)).isActive = true
-        leftButton!.leadingAnchor.constraint(equalTo: leadingAnchor, constant: CGFloat(Dimens.mediumMargin)).isActive = true
-        leftButton!.heightAnchor.constraint(equalToConstant: CGFloat(Dimens.button)).isActive = true
-        leftButton!.widthAnchor.constraint(greaterThanOrEqualToConstant: CGFloat(Dimens.minButtonWidth)).isActive = true
+        leftButton!.topAnchor.constraint(equalTo: topAnchor, constant: Dimens.mediumMargin).isActive = true
+        leftButton!.leadingAnchor.constraint(equalTo: leadingAnchor, constant: Dimens.mediumMargin).isActive = true
+        leftButton!.heightAnchor.constraint(equalToConstant: Dimens.button).isActive = true
+        leftButton!.widthAnchor.constraint(greaterThanOrEqualToConstant: Dimens.minButtonWidth).isActive = true
     }
     
     func setRightAction(title: String, action: @escaping () -> Void) {
@@ -56,10 +56,10 @@ class CheddarActionBar: UIView {
         rightButton!.title = title
         addSubview(rightButton!)
         rightButton!.translatesAutoresizingMaskIntoConstraints = false
-        rightButton!.topAnchor.constraint(equalTo: topAnchor, constant: CGFloat(Dimens.mediumMargin)).isActive = true
-        rightButton!.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -CGFloat(Dimens.mediumMargin)).isActive = true
-        rightButton!.heightAnchor.constraint(equalToConstant: CGFloat(Dimens.button)).isActive = true
-        rightButton!.widthAnchor.constraint(greaterThanOrEqualToConstant: CGFloat(Dimens.minButtonWidth)).isActive = true
+        rightButton!.topAnchor.constraint(equalTo: topAnchor, constant: Dimens.mediumMargin).isActive = true
+        rightButton!.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -Dimens.mediumMargin).isActive = true
+        rightButton!.heightAnchor.constraint(equalToConstant: Dimens.button).isActive = true
+        rightButton!.widthAnchor.constraint(greaterThanOrEqualToConstant: Dimens.minButtonWidth).isActive = true
     }
 
 }
